+++ lib/matlab/WB_NODE_MUSCLE.m
@@ -1,2 +1,0 @@
-function value = WB_NODE_MUSCLE
-value = 18;