+++ projects/robots/gctronic/e-puck/plugins/remote_controls/e-puck_wifi/DeviceManager.cpp
@@ -73,6 +73,9 @@ DeviceManager::DeviceManager() {
     mDevices.push_back(mLightSensors[i]);
   }
 
+  mTofSensor = new SingleValueSensor(wb_robot_get_device("tof"), 0);
+  mDevices.push_back(mTofSensor);
+
   mMotors[0] = new Motor(wb_robot_get_device("left wheel motor"), 0);
   mDevices.push_back(mMotors[0]);
   mMotors[1] = new Motor(wb_robot_get_device("right wheel motor"), 1);
@@ -143,6 +146,7 @@ void DeviceManager::clear() {
     mMotors[i] = NULL;
     mPositionSensors[i] = NULL;
   }
+  mTofSensor = NULL;
 }
 
 void DeviceManager::apply(int simulationTime) {