+++ resources/web/wwi/x3d_scene.js
@@ -12,6 +12,18 @@ class X3dScene { // eslint-disable-line no-unused-vars
     this.sceneModified = false;
     this.useNodeCache = {};
     this.objectsIdCache = {};
+
+    // The Mozilla WebGL implementation does not support automatic mimaps generation on float32 cube textures.
+    // - Warning (JS console): "Texture at base level is not unsized internal format or is not color-renderable or texture-filterable."
+    // - The related OpenGL specification is known as cryptic about this topic:
+    //   - https://www.khronos.org/registry/OpenGL-Refpages/es3/html/glGenerateMipmap.xhtml
+    //   - https://stackoverflow.com/questions/44754479/issue-with-rgba32f-texture-format-and-mipmapping-using-opengl-es-3-0
+    //   - https://stackoverflow.com/questions/56829454/unable-to-generate-mipmap-for-half-float-texture
+    const gl = document.createElement('canvas').getContext('webgl');
+    const glVendor = gl.getParameter(gl.VENDOR);
+    this.enableHDRReflections = glVendor !== 'Mozilla';
+    if (!this.enableHDRReflections)
+      console.warn('HDR reflections are not implemented for the current hardware.');
   }
 
   init(texturePathPrefix = '') {
@@ -52,6 +64,8 @@ class X3dScene { // eslint-disable-line no-unused-vars
     this.composer = new THREE.EffectComposer(this.renderer);
     let renderPass = new THREE.RenderPass(this.scene, this.viewpoint.camera);
     this.composer.addPass(renderPass);
+    this.bloomPass = new THREE.Bloom(new THREE.Vector2(window.innerWidth, window.innerHeight));
+    this.composer.addPass(this.bloomPass);
     this.hdrResolvePass = new THREE.ShaderPass(THREE.HDRResolveShader);
     this.composer.addPass(this.hdrResolvePass);
     var fxaaPass = new THREE.ShaderPass(THREE.FXAAShader);
@@ -72,13 +86,32 @@ class X3dScene { // eslint-disable-line no-unused-vars
   }
 
   render() {
+    // Set maximum rendering frequency.
+    // To avoid slowing down the simulation rendering the scene too often, the last rendering time is checked
+    // and the rendering is performed only at a given maximum frequency.
+    // To be sure that no rendering request is lost, a timeout is set.
+    const renderingMinTimeStep = 40; // Rendering maximum frequency: every 40 ms.
+    let currentTime = (new Date()).getTime();
+    if (this.nextRenderingTime && this.nextRenderingTime > currentTime) {
+      if (!this.renderingTimeout)
+        this.renderingTimeout = setTimeout(() => this.render(), this.nextRenderingTime - currentTime);
+      return;
+    }
+
+    // Apply pass uniforms.
     this.hdrResolvePass.material.uniforms['exposure'].value = 2.0 * this.viewpoint.camera.userData.exposure; // Factor empirically found to match the Webots rendering.
+    this.bloomPass.threshold = this.viewpoint.camera.userData.bloomThreshold;
+    this.bloomPass.enabled = this.bloomPass.threshold >= 0;
 
     if (typeof this.preRender === 'function')
       this.preRender(this.scene, this.viewpoint.camera);
     this.composer.render();
     if (typeof this.postRender === 'function')
       this.postRender(this.scene, this.viewpoint.camera);
+
+    this.nextRenderingTime = (new Date()).getTime() + renderingMinTimeStep;
+    clearTimeout(this.renderingTimeout);
+    this.renderingTimeout = null;
   }
 
   resize() {
@@ -109,8 +142,17 @@ class X3dScene { // eslint-disable-line no-unused-vars
     this.useNodeCache = {};
     this.root = undefined;
     this.scene.background = new THREE.Color(0, 0, 0);
+
+    /*
+    // Code to debug bloom passes.
+    var geometry = new THREE.PlaneGeometry(5, 5);
+    var material = new THREE.MeshStandardMaterial({color: 0xffffff, side: THREE.DoubleSide});
+    this.bloomPass.debugMaterial = material;
+    var plane = new THREE.Mesh(geometry, material);
+    this.scene.add(plane);
+    */
+
     this.onSceneUpdate();
-    this.render();
   }
 
   deleteObject(id) {
@@ -137,12 +179,12 @@ class X3dScene { // eslint-disable-line no-unused-vars
     if (object === this.root)
       this.root = undefined;
     this.onSceneUpdate();
-    this.render();
   }
 
   loadWorldFile(url, onLoad) {
     this.objectsIdCache = {};
     var loader = new THREE.X3DLoader(this);
+    loader.enableHDRReflections = this.enableHDRReflections;
     loader.load(url, (object3d) => {
       if (object3d.length > 0) {
         this.scene.add(object3d[0]);
@@ -173,11 +215,11 @@ class X3dScene { // eslint-disable-line no-unused-vars
     if (parentId && parentId !== 0)
       parentObject = this.getObjectById('n' + parentId);
     var loader = new THREE.X3DLoader(this);
-    var objects = loader.parse(x3dObject);
-    if (typeof parentObject !== 'undefined') {
-      objects.forEach((o) => { parentObject.add(o); });
+    loader.enableHDRReflections = this.enableHDRReflections;
+    var objects = loader.parse(x3dObject, parentObject);
+    if (typeof parentObject !== 'undefined')
       this._updateUseNodesIfNeeded(parentObject, parentObject.name.split(';'));
-    } else {
+    else {
       console.assert(objects.length <= 1 && typeof this.root === 'undefined'); // only one root object is supported
       objects.forEach((o) => { this.scene.add(o); });
       this.root = objects[0];
@@ -397,42 +439,6 @@ class X3dScene { // eslint-disable-line no-unused-vars
     return undefined;
   }
 
-  applyEquirectangularBackground(image) {
-    var texture;
-    if (image.data) {
-      texture = new THREE.DataTexture(image.data, image.width, image.height);
-      texture.encoding = THREE.RGBEEncoding;
-      texture.minFilter = THREE.NearestFilter;
-      texture.magFilter = THREE.NearestFilter;
-      texture.flipY = true;
-    } else {
-      texture = new THREE.Texture(image);
-      texture.minFilter = THREE.LinearFilter;
-      texture.magFilter = THREE.LinearFilter;
-    }
-    texture.needsUpdate = true;
-
-    var cubemapGenerator = new THREE.EquirectangularToCubeGenerator(texture, { resolution: image.width });
-    this.scene.background = cubemapGenerator.renderTarget;
-
-    var cubeMapTexture = cubemapGenerator.update(this.renderer);
-
-    var pmremGenerator = new THREE.PMREMGenerator(cubeMapTexture);
-    pmremGenerator.update(this.renderer);
-
-    var pmremCubeUVPacker = new THREE.PMREMCubeUVPacker(pmremGenerator.cubeLods);
-    pmremCubeUVPacker.update(this.renderer);
-
-    this.scene.background.userData = {};
-    this.scene.background.userData.isHDR = true;
-    this.scene.background.userData.texture = pmremCubeUVPacker.CubeUVRenderTarget.texture;
-    this._setupEnvironmentMap();
-
-    texture.dispose();
-    pmremGenerator.dispose();
-    pmremCubeUVPacker.dispose();
-  }
-
   // private functions
   _setupLights(directionalLights) {
     if (!this.root)
@@ -461,24 +467,26 @@ class X3dScene { // eslint-disable-line no-unused-vars
     var isHDR = false;
     var backgroundMap;
     if (this.scene.background) {
-      if (typeof this.scene.background.userData !== 'undefined' && this.scene.background.userData.isHDR) {
-        isHDR = true;
-        backgroundMap = this.scene.background.userData.texture;
-      } else if (this.scene.background.isCubeTexture)
-        backgroundMap = this.scene.background;
-      else if (this.scene.background.isColor) {
+      if (this.scene.background.isColor) {
         let color = this.scene.background.clone();
         color.convertLinearToSRGB();
         backgroundMap = TextureLoader.createColoredCubeTexture(color);
-      }
+      } else
+        backgroundMap = this.scene.background;
+    }
+
+    if (typeof this.scene.userData.irradiance !== 'undefined') {
+      isHDR = true;
+      backgroundMap = this.scene.userData.irradiance;
     }
 
     this.scene.traverse((child) => {
       if (child.isMesh && child.material && child.material.isMeshStandardMaterial) {
         var material = child.material;
         material.envMap = backgroundMap;
-        if (isHDR)
-          material.envMapIntensity = 0.2; // Factor empirically found to match the Webots rendering.
+        material.envMapIntensity = isHDR ? 0.6 : 1.0; // Factor empirically found to match the Webots rendering.
+        if (typeof this.scene.userData.luminosity !== 'undefined')
+          material.envMapIntensity *= this.scene.userData.luminosity;
         material.needsUpdate = true;
       }
     });